🚚 move `qfr_python` target into the main source tree and change it to a real library instead of header-only
